+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/Constants.kt
@@ -98,6 +98,8 @@ object Constants {
         const val INTELLIJ_PLATFORM_PLUGIN_LOCAL = "intellijPlatformPluginLocal"
         const val INTELLIJ_PLATFORM_PLUGIN_COMPOSED_MODULE = "intellijPlatformPluginComposedModule"
         const val INTELLIJ_PLATFORM_PLUGIN_MODULE = "intellijPlatformPluginModule"
+        const val INTELLIJ_PLATFORM_PLUGIN_ELEMENTS = "intellijPlatformPluginElements"
+        const val INTELLIJ_PLATFORM_PLUGIN_MODULE_DEPENDENCIES = "intellijPlatformPluginModuleDependencies"
         const val INTELLIJ_PLATFORM_PLUGIN = "intellijPlatformPlugin"
         const val INTELLIJ_PLATFORM_BUNDLED_PLUGINS = "intellijPlatformBundledPlugins"
         const val INTELLIJ_PLATFORM_BUNDLED_MODULES = "intellijPlatformBundledModules"